eyeclinc/static/requirements.txt
flask>=3.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
bcrypt>=4.1.0
openpyxl>=3.1.0
gunicorn>=21.2.0
